requirements.txt
alembic==1.0.5
atomicwrites==1.2.1
attrs==18.2.0
backports.shutil-get-terminal-size==1.0.0
Click==7.0
decorator==4.3.0
docopt==0.6.2
enum34==1.1.6
Flask==1.0.2
Flask-Migrate==2.3.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
funcsigs==1.0.2
ipdb==0.11
ipython==5.8.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
Jinja2==2.10
Mako==1.0.7
MarkupSafe==1.1.0
more-itertools==4.3.0
passlib==1.7.1
pathlib2==2.3.3
pexpect==4.6.0
pickleshare==0.7.5
pluggy==0.8.0
prompt-toolkit==1.0.15
ptyprocess==0.6.0
py==1.7.0
Pygments==2.3.0
pytest==4.0.2
python-dateutil==2.7.5
python-decouple==3.1
python-editor==1.0.3
scandir==1.9.0
simplegeneric==0.8.1
six==1.12.0
SQLAlchemy==1.2.15
SQLAlchemy-Utils==0.33.9
traitlets==4.3.2
wcwidth==0.1.7
Werkzeug==0.14.1
WTForms==2.2.1
